custom_agents/requirements.txt
langchain_ollama
duckduckgo-search
markdown-to-json
selenium
watchdog
langgraph-cli
python-frontmatter
jsonschema
pydash